chore: rename file `nginx-kind.ts` => `nginx-kind-manifests.ts`
