+++ core/src/plugins/kubernetes/local/nginx-kind-manifests.ts
